fix bug for closed splines, which worked with AutoCAD but not with BricsCAD
BUG: Autodesk docs say for closed spline n + 1 uniform knots,
which is invalid for BricsCAD and the basic spline
definition, (n + degree + 1) knots are required.

Solution was the common repetition of the first 'degree' control points,
and also weights for rational splines, at the end. This solution produces
the same result for AutoCAD and BricsCAD.
But closed rational splines are now different in AutoCAD and BricsCAD.