remove dependency from doc and entitydb for creating sub-entities for INSERT and POLYLINE
A doc.entitydb.refresh() call is done before DXF export to add new sub-entities
of INSERT and POLYLINE to the entity database, this call also can be done
manually at anytime.

This is only the case for adding sub-entities direct to INSERT or
POLYLINE entities, because DXF entities will not have direct access
to the DXF document and entity database in the future.

All polyline factory functions like add_polyline2d() will add the
new VERTEX entities automatically as in the past.