add synced_bytes_loader()
Does not raise DXFStructureError on invalid group codes,
instead skips lines until a valid group code or EOF is found.

This can remove invalid lines before group codes, but can not
detect invalid lines between group code and tag value.

This is a very aggressive method and it is unlikely that
files with such invalid DXF structures can be healed,
therefore this is not the default bytes loader.