fix bug for DXF R2018 - HATCH extrusion vector (210) is mandatory?
Example AutodeskProducts/Civil3D_2018.dxf was damaged for Trueview 2020,
because of missing the extrusion vector (0, 0, 1) for the HATCH entities,
the DXF reference also says the extrusion vector is optional.

Good News: TrueView ignores the re-encoded binary data in the XRECORD
entities (see previous commit), at least not the whole file is lost.