added adaptive recursive flattening for bezier curves
This algorithm tries to approximate e bezier curve to a given
maximum distance value by as few segments as possible, this
implementation uses a simple method to calculate this distance,
which may must be improved.

The next important task is to find the optimal distance value to get
a smooth curve for the path approximation in the the drawing
add-on.