add fonts.py module
A module to store font style properties beyond the
TTF file name in a backend agnostic way, at least for
commonly used fonts.

DXF stores only the TTF file name, which is not the
way backends choose a font, they need more information
to replace a font by similar font, if the original
font is not available on a system.