use CppVec3 in Evaluator.derivative
Creating a vector[CppVec3] was no gain of speed.

Current state of bspline C-extension:
10-12 faster than the pure Python implementation on CPython
2-4x faster than the warmed up pypy3 JIT

The important part, the point calculation without weights and
derivatives for clamped B-splines, is on the upper side of
the speed gain.

The algorithm seems to be broken for unclamped B-splines, but
these kind of B-splines are rare and the algorithm does not
crash!

clamped = starts at the first control point and ends at
the last control point.