remove proxy graphic support of ACAD_PROXY_OBJECT
This is not the proxy graphic, this is the proxy data.
No clue where the proxy graphic of application defined
entities is stored, should be ACAD_PROXY_ENTITY,
but this entity very seldom exist.

Goal was a basic support for "Map 3D" or "Civil 3D" entities
by converting the proxy graphic into DXF primitives.