add new implementation of the INSERT transformation
The new implementation solves issue #380, but I have
to admit, that this is just an trail&error approach,
I don't really know what I am doing here.

This implementation just transforms the source OCS to
detect the new axis, without applying the source
scaling and rotation.

Scaling and rotation will be transformed in a second step.
This is a much simpler approach than the previous
implementation and I am not sure if this works for
every situation.

Just one old test failed and it seems, that the new
implementation is more robust against transformation
errors.