use complex MTEXT rendering for columns or containing inline codes, closes #524
This is a good compromise: use the complex MTEXT render if there are columns
or inline formatting codes present, otherwise use the simple MTEXT renderer,
which is much faster. This avoids unnecessary long taking rendering
for MTEXT entities in DIMENSION entities.

Removed the option to choose the rendering method from the Frontend() class,
the "view" and the "draw" command.