reintroduce the sphinx_rtd_theme wide table fix
This may only work with the sphinx_rtd_theme < 1.0,
but there is an annoying bug in the theme, which disables
line wraps in tables. This fix overrides the default
CSS, but disables the whole theme when using the
sphinx_rtd_theme>=1.0.

This may break the docs at ezdxf.readthedocs.io