add emergency exit to Path.flattening() #574
Added a check for suddenly big distances (>1e12) which
can occur for very big coordinates.

2.9999999999e99 - 2.9999999998e99 = 9.999985256176552e+88

A small floating point imprecision in the mantissa can lead
to large differences, which broke the current implementation
of the path flattening algorithm.

The fix is to stop the recursion for sagitta distances > 1e12,
which is just an educated guess and is most likely not
the final solution for this issue.