return ACAD_TABLE content as virtual entities from associated anonymous block
Returning the content of the associated anonymous block seems to be
more reliable than the proxy graphic content.

The proxy graphic for the example "acad_table_with_blk_ref.dxf" shows
the block reference used in the table at an invalid location, maybe an
error in applying the transformation matrices stored in the proxy
graphic stream.