improve large coordinate handling for Bezier4P.flattening() #574 #663
This solution is also not perfect, the result differs from
the result with small coordinates, but at least no recursion
error will be raised.

The perfect solution would be:

1. detect curve control point extents
2. move curve near origin and store original offset
3. flatten curve and add offset to result vertices
4. move curve back, if curve processing handled inplace
   as the current implementation does

This is too much additional processing for a slightly
better result and large coordinates are the exception
like elevation > 1e12 in some 2D DXF files!