add type annotation for compiling by mypyc
This makes the module _mapbox_earcut.py compilable by mypyc,
but not inplace! Speed improvement by compiling by mypyc
about factor 4.
This is at least a quick test to show if a Cython
implementation is worth the effort.

The inplace compilation shows a cyclic import error because
there is an implicit import of ezdxf.math.__init__.py done,
which is not needed by _mapbox_earcut.py. This maybe can be
fixed by a mypyc option or by a different file structure.