divide the Recorder class into Recorder- and Player class
The Recorder class cannot modify the recordings.
The Player class cannot extend the recordings.

Recorder.copy_player() creates a Player class with
a copy of the recordings. (needs more memory - but safe)

Recorder.shared_player() creates a Player class with
the original recordings which modifies the original
recordings at transformation!!! (faster and lighter - but dangerous)