fix #907
The version tag in ATTDEF and ATTRIB has the same group code
as the lock_position tag and the implementation required both
to be present to work properly.

The new implementation uses the last occurrence of group code
280 as lock position tag. The version tag is always 0, so if
lock_position is missing and the version tag is present,
lock_position is set to 0, which is the default value.