add special exception for system under test #921
The fonts module has a special font cache for the system
under test (SUT). This is a problem for users of the
library, who want to test ezdxf function related to font
rendering.

The current solution is to raise a special exception
SystemUnderTest() with a link to the issue #921, which
show a solution for the problem.